+++ src/pages/Review/components/CommentItem.jsx
@@ -1,30 +1,17 @@
 import React, { useState } from "react";
 
-//components
+// Components
 import DeleteModal from "./DeleteModal";
 
-//hooks
+// Hooks
 import useMediaQueries from "../../../hooks/useMediaQueries";
 
-//styles
-import trash from "../assets/trash.svg";
-import {
-  CommentBox,
-  CommentHeader,
-  CommentNumber,
-  CommentText,
-  CommentContent,
-  CommentFooter,
-  CommentDate,
-  CommentActions,
-  DeleteIconContainer,
-  DeleteText,
-  CommentDivider,
-  ImageWrapper,
-} from "../components/CommentStyle";
+// Styles
+import * as S from "../components/CommentStyle";
 
 const Comment = ({ comment }) => {
   const { isMobile, isTablet, isDesktop } = useMediaQueries();
+  const mediaUrl = import.meta.env.VITE_MEDIA_URL;
 
   // 삭제 모달 로직
   const [isModalOpen, setIsModalOpen] = useState(false);
@@ -43,71 +30,75 @@ const Comment = ({ comment }) => {
   };
 
   return (
-    <CommentBox
+    <S.CommentBox
       $isMobile={isMobile}
       $isTablet={isTablet}
       $isDesktop={isDesktop}
     >
-      <CommentHeader>
-        <CommentNumber
+      <S.CommentHeader>
+        <S.CommentNumber
           $isMobile={isMobile}
           $isTablet={isTablet}
           $isDesktop={isDesktop}
         >
           {comment.id}
-        </CommentNumber>
-      </CommentHeader>
-      <CommentText
+        </S.CommentNumber>
+      </S.CommentHeader>
+      <S.CommentText
         $isMobile={isMobile}
         $isTablet={isTablet}
         $isDesktop={isDesktop}
       >
         {comment.text}
-      </CommentText>
+      </S.CommentText>
 
       {/* 이미지가 있을 때만 보여주기 */}
       {comment.image && (
-        <CommentContent>
-          <ImageWrapper
+        <S.CommentContent>
+          <S.ImageWrapper
             $isMobile={isMobile}
             $isTablet={isTablet}
             $isDesktop={isDesktop}
             image={comment.image}
-          >
-            {/* <ImagePlaceholder>+</ImagePlaceholder> */}
-          </ImageWrapper>
-        </CommentContent>
+          ></S.ImageWrapper>
+        </S.CommentContent>
       )}
 
-      <CommentFooter>
-        <CommentDate
+      <S.CommentFooter>
+        <S.CommentDate
           $isMobile={isMobile}
           $isTablet={isTablet}
           $isDesktop={isDesktop}
         >
           {comment.date}
-        </CommentDate>
-        <CommentActions>
-          <DeleteIconContainer onClick={handleDeleteClick}>
-            <img src={trash} alt="Delete Icon" width="17" height="17" />
-          </DeleteIconContainer>
-          <DeleteText
+        </S.CommentDate>
+        <S.CommentActions>
+          <S.DeleteIconContainer onClick={handleDeleteClick}>
+            <img
+              src={`${mediaUrl}Review/trash.png`}
+              alt="Delete Icon"
+              width="17"
+              height="17"
+            />
+          </S.DeleteIconContainer>
+          <S.DeleteText
+            onClick={handleDeleteClick}
             $isMobile={isMobile}
             $isTablet={isTablet}
             $isDesktop={isDesktop}
           >
             삭제
-          </DeleteText>
-        </CommentActions>
-      </CommentFooter>
-      <CommentDivider />
+          </S.DeleteText>
+        </S.CommentActions>
+      </S.CommentFooter>
+      <S.CommentDivider />
       {isModalOpen && (
         <DeleteModal
           onClose={handleCloseModal}
           onConfirm={handleConfirmDelete}
         />
       )}
-    </CommentBox>
+    </S.CommentBox>
   );
 };
