+++ src/pages/Review/components/CommentStyle.jsx
@@ -1,14 +1,42 @@
 import styled from "styled-components";
 import useMediaQueries from "../../../hooks/useMediaQueries";
 
-// 댓글 전체 컨테이너
+// CommentSection 스타일
+
 export const CommentsContainer = styled.div`
   width: 100%;
   display: flex;
   flex-direction: column;
   gap: ${(props) =>
-    props.isMobile ? "20px" : props.isTablet ? "12px" : "24px"};
-  padding: ${(props) => (props.isMobile ? "20px" : "24px")};
+    props.$isMobile ? "20px" : props.$isTablet ? "12px" : "24px"};
+  padding: ${(props) => (props.$isMobile ? "20px" : "24px")};
+`;
+
+export const PaginationContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  gap: 8px;
+`;
+
+export const PageButton = styled.button`
+  padding: 8px 12px;
+  border: none;
+  border-radius: 8px;
+  background-color: white;
+  color: #0a0b0a;
+  font-size: 16px;
+  font-family: "Inter", sans-serif;
+  font-weight: 400;
+  line-height: 16px;
+  cursor: pointer;
+
+  &:hover {
+    background-color: #e7ebef;
+  }
+`;
+
+export const CurrentPageButton = styled(PageButton)`
+  background-color: #e7ebef;
 `;
 
 // CommentItem 스타일
@@ -35,10 +63,9 @@ export const CommentHeader = styled.div`
 export const CommentNumber = styled.div`
   color: #4a5e6d;
   font-size: ${(props) => (props.$isMobile ? "14px" : "20px")};
-  font-family: "MonSemiBold"
+  font-family: "MonSemiBold";
   font-weight: 400;
   line-height: 30px;
-  font-size: ${(props) => (props.$isMobile ? "14px" : "20px")};
   word-wrap: break-word;
 `;
 
@@ -70,7 +97,6 @@ export const CommentDate = styled.div`
   font-family: "MonRegular", sans-serif;
   font-weight: 400;
   line-height: ${(props) => (props.$isMobile ? "18px" : "24px")};
-
   word-wrap: break-word;
 `;
 
@@ -82,14 +108,15 @@ export const CommentActions = styled.div`
   justify-content: center;
   align-items: center;
   gap: 10px;
-  cursor: pointer; 
+  cursor: pointer;
+
   &:hover {
-    opacity: 0.8; 
+    opacity: 0.8;
+  }
 `;
 
 export const DeleteIconContainer = styled.div`
   display: flex;
-  }
 `;
 
 export const DeleteText = styled.div`
@@ -128,16 +155,6 @@ export const ImageWrapper = styled.div`
   }
 `;
 
-//후기 input 필드로 옮기기
-// export const ImagePlaceholder = styled.div`
-//   color: #ced7de;
-//   font-size: 24px;
-//   font-family: "Noto Sans KR", sans-serif;
-//   font-weight: 400;
-//   line-height: 38.83px;
-//   word-wrap: break-word;
-// `;
-
 export const CommentContent = styled.div`
   display: inline-flex;
   justify-content: flex-start;
@@ -145,29 +162,132 @@ export const CommentContent = styled.div`
   gap: 20px;
 `;
 
-export const PaginationContainer = styled.div`
+// DeleteModal 스타일
+
+export const ModalWrapper = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
   display: flex;
   justify-content: center;
-  gap: 8px;
+  align-items: center;
+  z-index: 1000;
 `;
 
-export const PageButton = styled.button`
-  padding: 8px 12px;
+export const ModalContent = styled.div`
+  width: ${(props) => (props.$isMobile ? "340px" : "564px")};
+  height: ${(props) => (props.$isMobile ? "365.59px" : "442px")};
+  padding: ${(props) => (props.$isMobile ? "10px 20px" : "48px 80px")};
+  background: white;
+  box-shadow: 0px 8px 8px rgba(0, 0, 0, 0.25);
+  border-radius: 8px;
+`;
+
+export const ModalBox = styled.div`
+  padding: ${(props) => (props.$isMobile ? "48px 20px" : "16px 0 48px")};
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: ${(props) => (props.$isMobile ? "20px" : "40px")};
+`;
+
+export const ModalText = styled.div`
+  text-align: center;
+  color: #474747;
+  font-size: ${(props) => (props.$isMobile ? "18px" : "27px")};
+  font-family: "MonRegular";
+  font-weight: 400;
+  line-height: 38px;
+  overflow-wrap: break-word;
+`;
+
+export const PasswordInputWrapper = styled.div`
+  display: flex;
+  width: ${(props) => (props.$isMobile ? "300px" : "320px")};
+  height: ${(props) => (props.$isMobile ? "96px" : "96px")};
+  justify-content: space-between;
+  align-items: center;
+  gap: 10px;
+  padding: 24px 20px;
+  background: #fafafa;
+  border-radius: 16px;
+`;
+
+export const PasswordInput = styled.input`
+  width: 100%;
+  padding: 0px 150px 0px 0px;
+  color: #474747;
+  font-size: 18px;
+  font-family: "MonRegular";
+  line-height: 24px;
   border: none;
   border-radius: 8px;
-  background-color: white;
-  color: #0a0b0a;
-  font-size: 16px;
-  font-family: "Inter", sans-serif;
-  font-weight: 400;
-  line-height: 16px;
+  text-align: center;
+  outline: none;
+`;
+
+export const PasswordIcon = styled.div`
+  width: ${(props) => (props.$isMobile ? "32px" : "48px")};
+  height: ${(props) => (props.$isMobile ? "32px" : "48px")};
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+`;
+
+export const PasswordIconImage = styled.img`
+  width: ${(props) => (props.$isMobile ? "32px" : "48px")};
+  height: auto;
+`;
+
+export const ModalActions = styled.div`
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  padding: ${(props) => (props.$isMobile ? "0px 0px 10px 0px" : "0 12px")};
+`;
+
+export const CancelButton = styled.button`
+  flex: 1;
+  height: 45px;
+  background: white;
+  border: 1px solid #637d92;
+  border-radius: 100px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 14px;
+  font-family: "MonRegular";
+  color: #637d92;
   cursor: pointer;
 
   &:hover {
-    background-color: #e7ebef;
+    background-color: #637d92;
+    color: #ffffff;
   }
 `;
 
-export const CurrentPageButton = styled(PageButton)`
-  background-color: #e7ebef;
+export const ConfirmButton = styled.button`
+  flex: 1;
+  height: 45px;
+  background: white;
+  border: 1px solid #0068ff;
+  border-radius: 100px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 14px;
+  font-family: "MonRegular";
+  color: #0068ff;
+  cursor: pointer;
+  margin-left: 12px;
+
+  &:hover {
+    background-color: #0068ff;
+    color: #ffffff;
+  }
 `;
